offloading cleanups to sed(1) commands
More offloading to the send commands for cleanup.
basically the gawk ERE was matching even lines with spaces in the left and in the right, removed that as part of the RE
and moved that to the sed commands executed before calling awk. now the gawk RE is more compact and simple than before.